fix recommended changes
