--- src/components/FontSizes/Font.style.scss
@@ -1,8 +1,0 @@
-@import "../../styles/variables";
-
-
-@each $name, $entry in $sizes {
-  .size__#{$name} {
-    font-size: map-get($entry, "font");
-  }
-}
