+++ src/components/badge/Badge.style.scss
@@ -6,13 +6,13 @@
 
   padding: .25rem;
   display: inline-flex;
-  font-size: map-get($xs, "font");
+  font-size: $xs;
   width: fit-content;
 }
 
 @each $name, $color in $variants {
   .badge--#{$name} {
     @include opacityBox(false, $color);
-    font-size: map-get($xs, "font");
+    font-size: $xs;
   }
 }
