+++ src/components/card/Card.tsx
@@ -5,7 +5,7 @@ import {CardHeader, CardHeaderType} from "./CardHeader";
 import {CardTitle, CardTitleType} from "./CardTitle";
 import {CardSubTitle, CardSubTitleType} from "./CardSubTitle";
 import {CardFooter, CardFooterType} from "./CardFooter";
-import {Color} from "../../utils/utils";
+import {Color} from "../../utils/types";
 
 export type CardChildType = CardHeaderType | CardImgStyle | CardTitleType | CardSubTitleType | CardFooterType | any
