+++ src/components/menu/Menu.style.scss
@@ -55,7 +55,7 @@
   }
 
   &__section-title {
-    font-size: map-get($xs, "font");
+    font-size: $xs;
     color: rgba($white, .25);
     display: block;
     margin: .25rem 0 .25rem .25rem;
