+++ src/utils/types.ts
@@ -1,0 +1,7 @@
+export type Color = "primary" | "secondary" | "info" | "success" | "warning" | "error";
+
+export const Colors: Color[] = ["primary", "secondary", "info", "success", "warning", "error"]
+
+export type Size = "xs" | "sm" | "md" | "lg" | "xl";
+
+export const Sizes: Size[] = ["xs", "sm", "md", "lg", "xl"]
