+++ src/utils/utils.ts
@@ -1,13 +1,5 @@
 import React, {ReactNode, useEffect, useMemo, useState} from "react";
 
-export type Color = "primary" | "secondary" | "info" | "success" | "warning" | "error";
-
-export const Colors: Color[] = ["primary", "secondary", "info", "success", "warning", "error"]
-
-export type Size = "xs" | "sm" | "md" | "lg" | "xl";
-
-export const Sizes: Size[] = ["xs", "sm", "md", "lg", "xl"]
-
 
 export const getChild = (children: ReactNode | ReactNode[], child: React.FC<any>, required?: boolean): React.ReactElement | undefined => {
